requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0
trl>=0.4.0
accelerate>=0.20.0
datasets>=2.12.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
PyYAML>=6.0
wandb>=0.15.0
mlflow>=2.4.0
pycocotools>=2.0.6
evaluate>=0.4.0
nltk>=3.8
rouge-score>=0.1.2